Java sliver (using recursive)

+++ BOJ/Bronze/S27433.java
@@ -2,7 +2,7 @@ import java.util.Scanner;
 
 public class S27433
 {
-    public static int factorial(int n)
+    public static long factorial(int n)
     {
         if(n == 0 || n == 1) return 1;
         else
